Fixed db errors

+++ database/migrations/2020_03_05_000000_create_comments_table.php
@@ -16,7 +16,7 @@ class CreateCommentsTable extends Migration
         Schema::create('comments', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->text('description');
-            $table->integer('product_id');
+            $table->bigInteger('product_id');
             $table->foreign('product_id')->references('id')->on('products');
             $table->timestamps();
         });
